+++ CHANGELOG.md
@@ -9,6 +9,16 @@
 
 ### Fixed
 
+## [1.1.0] - 2022-08-10
+### Added
+- Add debug agent information
+- Add list open files function
+- Add list open ports function
+- Add more process name to icons mappings
+
+### Changed
+- Collect processes during everytime the tool window gets opened
+
 ## [1.0.1] - 2022-07-07
 ### Fixed
 - Prevent a NPE if the user, group or working directory of a process is unknown.
@@ -17,3 +27,4 @@
 ## [1.0.0] - 2022-07-02
 ### Added
 - Initial Release
+